Handle API language select with header instead of query param (#20)
* Use optional configuration block for getRequest()

Both getRequest() and getRequestAuth() add the same types of headers and logging functionality, with the auth version just adding authentication details. By adding an optional configuration callback to the base function, this duplication can be avoided and the authentication can just be added before the request gets send out.

* Add optional language parameter to getRequest() and getRequestAuth()

The language parameter is used to set the value of the "Accept-Language" header. It's always being added, defaulting to the client's default language, which in turn defaults to english.

* Update internal API clients to use the new language parameter for requests

* Update public MapInfoClient to use updated way of specifying the language

* Add Accept-Language verification capability

* Fix tests to properly verify the new language header instead of query param

* Add config block to getRequestAuth() as well