Update internal Http clients to use the new ApiLanguage enum
It's still possible to define a default language when constructing the client, making the client always send the Accept-Language header. If no language is defined though, the client will no longer send the header, therefore letting the API determine the language. This should default to english everytime, according to the docs.