Use Enums for language handling (#22)
* Add ApiLanguage enum

Every enum entry is backed by a Locale object. This locale object is also used to determine the language string that gets sent to the API. It's also possible to get the correct enum entry by providing either a Locale object or an ISO 639-1 langauge code.

* Update internal Http clients to use the new ApiLanguage enum

It's still possible to define a default language when constructing the client, making the client always send the Accept-Language header. If no language is defined though, the client will no longer send the header, therefore letting the API determine the language. This should default to english everytime, according to the docs.

* Update internal API clients to accept the new ApiLanguage enum

It's now also possible to pass null to the methods of the internal clients, making the http client not send any language specification at all.

* Update public API clients to optionally accept the new ApiLanguage enum

* Add @JvmOverloads to public API client methods that have default values

This improves overall JVM interoperability, especially since the tests are still in Groovy for now. This does increase the compiled size a bit as additional overloads get generated, but it's still a good thing imo.

* Update tests to reflect language changes

* Remove ability to define a default language when creating a client

* Bump version to 1.6

As this includes a breaking change in method signatures, we should release it as a new version.

* Cleanup comments and newlines