Switch from Gson to kotlinx.serialization (#30)
* Start linking appropriate endpoint documentation

The GW2 Wiki provides documentation about the various endpoints. Using KDoc's markdown feature, we can easily link to the appropriate Wiki URL that contains the documentation.

* Fix wrong WeaponSkillSlot Mapping

* Prepare data models for new serialization library

kotlinx.serialization sadly requires all data models to be explicitly annotated...

* Disable details property of Item model

See #28 for more details, will need to be correctly implemented later

* Update serializers for use with kotlinx.serialization

* Update internal HttpClient to be compatible with kotlinx.serialization

* Update tests to no longer use Gson

Admittedly, using kotlinx.serialization from Groovy is kinda ugly. Since there are plans to migrate to a fully Kotlin based stack, this shouldn't be a problem in the future though.

* Fix Profession, Skill and Trait data models by providing polymorphic hierarchy

The previous implementation simply dropped most of the unknown keys, losing information in the process. The serialization library provides support for polymorphic serialization, and since the API always provides a "type" attribute, we can just map it to our Fact subtypes.

I've also fixed several typos in property names and completed enums that weren't exhaustive.

* Fix Account data models

Just added default values here for optional properties

* Update tests for SkillsClient and TraitsClient

* Fix Achievement data models

Fixed typos, changed data types of numeric IDs to Int, and added the missing data model for Achievement Bits.

* Fix Character data models

Added default values to optional parameters and made certain properties nullable to adhere to API specs.

* Fix Guild data model

Added default values to optional properties and added subtypes for the different types of GuildLog types.

* Fix Item data model

Due to a lack of reflection, the serialization library can't deserialize values to Any. This is being addressed in #29 and needs to be fixed later. For now, access to the details is still manually possible by providing a JsonObject.

* Add new Quaggan data class to capture all properties retuend by the API

The API returns the picture URL as "url", so the Icon class can't be easily used here.

* Fix MapInfo data models

Added default values to optional properties and added missing Serializable annotations

* Fixed remaining data models

Just added missing properties or added default values to optional ones.

* Fixed custom serializers

PairSerializer can't be used as a delegate serializer here, because it's kind is set to CLASS. The API returns JSON arrays though, so we need something with kind of LIST.

* Make Json parser lenient

The mapping in TraitTier doesn't work without a lenient parser, because it's expecting a string while the API responds with a number literal.

* Fix WvW data models

* Remove redundant Gson version

We don't use Gson anymore.

* Fix typo in class name of Skill/Trait fact

* Use SerialDescriptor from delegated serializer in custom serializers

* Add missing documentation to new data classes

* update branch with newest master, increment project version

---------

Co-authored-by: Krzysztof Prajs <krzysztof.prajs@falest.pl>